requirements.txt
# AI Takashi API Server Requirements
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
cryptography>=41.0.0

# AI Takashi Client Requirements
PyQt5>=5.15.9
requests>=2.31.0

# Optional: Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
